refactor: move release publish pipeline into release app
