+++ apps/core/views/reports/release_publish/state/context.py
@@ -3,7 +3,7 @@
 Legacy import path: ``release_publish.state.context``.
 """
 
-from ..context import (  # noqa: F401
+from apps.release.publishing.context import (  # noqa: F401
     load_release_context,
     persist_release_context,
     sanitize_release_context,
