+++ apps/release/tests/test_publish_steps.py
@@ -2,10 +2,12 @@
 
 from __future__ import annotations
 
+from apps.core.views.reports.release_publish import pipeline as UI_PIPELINE
 from apps.core.views.reports.release_publish.views import (
     PUBLISH_STEPS as UI_PUBLISH_STEPS,
 )
 from apps.release.domain import PUBLISH_STEPS as DOMAIN_PUBLISH_STEPS
+from apps.release.publishing import pipeline as RELEASE_PIPELINE
 from apps.release.release_workflow import _build_release_workflow
 
 EXPECTED_STEP_ORDER = [
@@ -27,6 +29,11 @@ EXPECTED_STEP_ORDER = [
 def test_release_publish_steps_share_canonical_order() -> None:
     assert [name for name, _handler in DOMAIN_PUBLISH_STEPS] == EXPECTED_STEP_ORDER
     assert [name for name, _handler in UI_PUBLISH_STEPS] == EXPECTED_STEP_ORDER
+
+
+def test_core_release_publish_pipeline_path_is_adapter() -> None:
+    assert UI_PIPELINE is RELEASE_PIPELINE
+    assert UI_PUBLISH_STEPS is RELEASE_PIPELINE.PUBLISH_STEPS
 
 
 def test_headless_release_workflow_uses_canonical_order() -> None:
